v1.0.1: profile delete/nickname, USSD, phonebook, SMS export, logs, update check, unit tests
- Core: DeleteProfileAsync (BF33, no A0 wrapper), SetNicknameAsync (BF29 by ICCID),
  CusdService (GSM7/UCS2 auto-decode), PhonebookService (UCS2 names),
  SmsExportService (CSV/JSON), AppLog (7-day rolling), UpdateCheckService
- Fix: GSM7 decoder carry truncation (byte accumulator -> int) — roundtrip bug
- UI: profile enable/disable/rename/delete actions, USSD card, phonebook card,
  SMS export buttons, update check in Settings, global exception logging
- Tests: 30 unit tests (GSM7 standard vectors incl. 3GPP hellohello, ES10c
  byte-exact encoding vs lpac, BCD nibble order, export escaping)
- CI: GitHub Actions release workflow (test + double-config publish + assets)
- Docs: README roadmap (SM-DP+ download deferred with rationale), README.en.md

+++ src/CellPort.Core/At/SmsCodec.cs
@@ -72,7 +72,7 @@ public static class SmsCodec
     {
         var count = septetCount > 0 ? septetCount : data.Length * 8 / 7;
         var sb = new StringBuilder(count);
-        byte accumulator = 0;
+        int accumulator = 0;
         int bits = 0;
         int index = 0;
 
@@ -84,7 +84,7 @@ public static class SmsCodec
                 {
                     break;
                 }
-                accumulator |= (byte)(data[index++] << bits);
+                accumulator |= data[index++] << bits;
                 bits += 8;
             }
